* deleted function of renaming interviews * added alert dialog when trying to delete an interview

+++ src/main/java/Components/InterviewTree/Controller/InterviewTreeItemController.java
@@ -1,19 +1,15 @@
 package Components.InterviewTree.Controller;
 
-import Components.InterviewTree.Cell.Model.InterviewElement;
 import Components.InterviewTree.Cell.Model.InterviewItem;
-import Components.InterviewTree.Commands.DeleteItemPluggable;
-import Components.SchemaTree.Cell.Controllers.SchemaTreeCellController;
-import Components.SchemaTree.Cell.Models.SchemaProperty;
+
 import application.Configuration.Configuration;
 import application.History.HistoryManager;
-import javafx.event.EventHandler;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.MenuItem;
-import javafx.scene.input.MouseEvent;
 import utils.Removable.Commands.DeleteRemovableCommand;
-import utils.ResourceLoader;
-
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 
@@ -31,8 +27,18 @@ public class InterviewTreeItemController extends InterviewTreeCellController {
         super.initialize(url, resourceBundle);
         MenuItem deleteButton = new MenuItem(Configuration.langBundle.getString("delete"));
         deleteButton.setOnAction(actionEvent -> {
-            HistoryManager.addCommand(new DeleteRemovableCommand(element), true);
+            if (confirmDelete())
+                HistoryManager.addCommand(new DeleteRemovableCommand(element), true);
         });
         optionsMenu.getItems().add(deleteButton);
+    }
+
+    private boolean confirmDelete(){
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle(Configuration.langBundle.getString("delete_warning"));
+        alert.setHeaderText(Configuration.langBundle.getString("delete_interview"));
+        alert.setContentText(Configuration.langBundle.getString("delete_interview_text_alert"));
+        Optional<ButtonType> result = alert.showAndWait();
+        return result.get() == ButtonType.OK;
     }
 }
